Fixed a few things

+++ src/lauxlib.c
@@ -760,7 +760,7 @@ LUALIB_API const char *luaL_tolstring (lua_State *L, int idx, size_t *len) {
         break;
       default:
         lua_pushfstring(L, "%s: %p", luaL_typename(L, idx),
-                                            lua_topointer(L, idx));
+                                            (ptrdiff_t)lua_topointer(L, idx));
         break;
     }
   }
